fix(DFC-917): refactor and fix for navigation event not firing on method switch
