+++ packages/frontend-analytics/src/analytics/formResponseTracker/formResponseTracker.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, jest, test, beforeEach } from "@jest/globals";
 import { FormResponseTracker } from "./formResponseTracker";
 import { FormEventInterface } from "../formTracker/formTracker.interface";
-import * as pushToDataLayer from "../../utils/pushToDataLayer";
+import * as pushToDataLayer from "../../utils/pushToDataLayerUtil/pushToDataLayer";
 
 window.DI = { analyticsGa4: { cookie: { consent: true } } };
