+++ packages/frontend-analytics/src/analytics/selectContentTracker/selectContentTracker.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, jest, test, beforeEach } from "@jest/globals";
 import { SelectContentTracker } from "./selectContentTracker";
 import { SelectContentEventInterface } from "./selectContentTracker.interface";
-import * as pushToDataLayer from "../../utils/pushToDataLayer";
+import * as pushToDataLayer from "../../utils/pushToDataLayerUtil/pushToDataLayer";
 
 describe("selectContentTracker", () => {
   let newInstance: SelectContentTracker;
