+++ packages/frontend-analytics/src/utils/applyDefaultsUtil/applyDefaults.ts
