+++ packages/frontend-analytics/src/utils/dataScrapersUtils/dataScrapers.test.ts
