+++ packages/frontend-analytics/src/utils/dataScrapersUtils/dataScrapers.ts
