+++ packages/frontend-analytics/src/utils/validateParameterUtils/validateParameter.ts
@@ -2,7 +2,7 @@
  *   Takes a paremeter passed in from external use of a tracker and asserts it is defined, is a string and has a length allowed by GA
  */
 
-import { stripPIIFromString } from "./piiRemover";
+import { stripPIIFromString } from "../piiRemoverUtil/piiRemover";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function validateParameter(parameter: any, maxLength: number) {
